Fix bug in ezANOVA to ensure type=3 works properly
Apparently, to get proper type-III results from car::Anova, one has to use contrasts that are orthogonal to the intercept, which might not have been set by the user. Failing to do so will yield erroneous results for within effects that do not involve a between effect. (... but the "erroneous" values will be close to those obtained from type=2, which you should be using anyway!)