erRound-Function
Hey there,

I wrote a little function to round the results of ezANOVA objects (i.e.
the list format). I’d like to contribute it. If you like it, I’d move
on and write a more flexible version that is compatible with all the
lists ez creates.
Cheers,
Florian